+++ src/commands/checkout.js
@@ -1,7 +1,7 @@
 const runAliasCommand = require('./tools/runAliasCommand');
 
 runAliasCommand({
-    description: 'Alias of "hulky -- git checkout"',
+    description: 'Alias of "mayan -- git checkout"',
     asynchronous: true,
     command: args => `git checkout ${args.join(' ')}`,
 }).then().catch(e => { console.error(e); process.exit(1); });