+++ src/commands/tools/config.js
@@ -4,9 +4,9 @@ const { promisify } = require('util');
 
 const readFile = promisify(fs.readFile);
 
-// Recursively find the .hulkyrc.js file in the current directory or its parents
+// Recursively find the .mayan.js file in the current directory or its parents
 function findHulkyRc(dir) {
-    const filePath = resolve(dir, '.hulkyrc.js');
+    const filePath = resolve(dir, '.mayan.js');
 
     if (fs.existsSync(filePath)) {
         return filePath;
@@ -25,7 +25,7 @@ const filePath = findHulkyRc(process.cwd());
 
 if (!filePath) {
     // eslint-disable-next-line no-console
-    console.error('No .hulkyrc.js file found in the current directory or its parents');
+    console.error('No .mayan.js file found in the current directory or its parents');
     process.exit(1);
 }