+++ src/commands/tools/forEachProject.js
@@ -9,7 +9,7 @@ const { bgYellow, black } = chalk;
 let config;
 
 function warnMissing(path) {
-    warn(bgYellow(black('WARN')), `${path} does not exist. Run 'hulky init' to initialize uncloned repositories.`);
+    warn(bgYellow(black('WARN')), `${path} does not exist. Run 'mayan init' to initialize uncloned repositories.`);
 }
 
 module.exports = async function forEachProject({