+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## [0.2] - 2022-01-31
+## [0.3] - 2022-01-31
 ### Breaking
-- Package name `@legion-workspace/cli` and command name `legion` both changed to `hulky`
-- `.legionrc.js` is now `.hulkyrc.js`
+- Package name `@legion-workspace/cli` and command `legion` both changed to `mayan`
+- `.legionrc.js` is now `.mayan.js`
+